Split layer and upload from repository
Layer upload moved to its own file with its own unit tests
